# docs/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "خالد سعد — عقد api/v1",
        "version": "1.0.0",
        "description": "مولَّدة من جدول المسارات — لا تُحرَّر يدويًّا.\n\nإصدارٌ واحد لا ثانٍ: يتطوّر api/v1 في مكانه، محروسًا ببوابة\nحدّ أدنى لإصدار التطبيق تُشحن قبل أي تغيير في العقد."
    },
    "servers": [
        {
            "url": "http://localhost/khaledsaad/public/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "kind",
                            "code",
                            "title",
                            "message",
                            "user_action"
                        ],
                        "properties": {
                            "kind": {
                                "type": "string",
                                "enum": [
                                    "ours",
                                    "theirs",
                                    "input"
                                ]
                            },
                            "code": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "user_action": {
                                "nullable": true,
                                "description": "يجب أن يكون null حين kind = ours.",
                                "type": "object",
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "route": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            },
                            "retry_after": {
                                "type": "integer",
                                "nullable": true
                            }
                        }
                    }
                }
            },
            "Envelope": {
                "type": "object",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object",
                        "properties": {
                            "locale": {
                                "type": "string"
                            },
                            "server_time": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "Success": {
                "description": "نجاح",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Envelope"
                        }
                    }
                }
            },
            "Error": {
                "description": "عطل مصنَّف",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    },
    "paths": {
        "/api/v1/admin/catalog": {
            "get": {
                "operationId": "api.v1.admin.catalog",
                "summary": "api.v1.admin.catalog",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/public/runs/{run}/preflight": {
            "get": {
                "operationId": "api.v1.public.runs.preflight",
                "summary": "api.v1.public.runs.preflight",
                "tags": [
                    "public"
                ],
                "security": [],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/runs/{run}/insights": {
            "post": {
                "operationId": "api.v1.runs.insights",
                "summary": "api.v1.runs.insights",
                "tags": [
                    "runs"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/runs/{run}/preflight": {
            "get": {
                "operationId": "api.v1.runs.preflight",
                "summary": "api.v1.runs.preflight",
                "tags": [
                    "runs"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    }
}